+++ crypto-sodium/test/Test/Crypto/Sign.hs
@@ -5,11 +5,11 @@
 -- | “Integration” tests: using Sign with our helpers.
 module Test.Crypto.Sign where
 
-import Hedgehog (Property, forAll, property, tripping)
+import Hedgehog (Property, evalMaybe, forAll, property, tripping)
 import Hedgehog.Internal.Property (forAllT)
 
 import Control.Monad.IO.Class (liftIO)
-import Data.ByteArray.Sized (unsafeSizedByteArray)
+import Data.ByteArray.Sized (sizedByteArray)
 import Data.ByteString (ByteString)
 
 import qualified Hedgehog.Gen as G
@@ -21,12 +21,12 @@ import qualified Crypto.Sign as Sign
 
 
 seedSize :: R.Range Int
-seedSize = R.singleton $ fromIntegral Na.crypto_box_seedbytes
+seedSize = R.singleton $ fromIntegral Na.crypto_sign_seedbytes
 
 
 hprop_encode_decode_seed :: Property
 hprop_encode_decode_seed = property $ do
-    seed <- fmap unsafeSizedByteArray . forAll $ G.bytes seedSize
+    seed <- evalMaybe . sizedByteArray =<< forAll (G.bytes seedSize)
     (pk, sk) <- forAllT . liftIO $ Sign.keypairFromSeed seed
     msg <- forAll $ G.bytes (R.linear 0 1_000)
     tripping msg (encodeBs sk) (decodeBs pk)